% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10to11.R
\name{icd10to11}
\alias{icd10to11}
\title{Convert ICD10 codes to ICD11}
\usage{
icd10to11(x, ...)
}
\arguments{
\item{x}{character. ICD10 codes. If you supply only the start of the ICD10 code, it will match based on the start of the code.}

\item{...}{arguments passed to \code{icd11_lookup}}
}
\value{
a list.
}
\description{
This function used a mapping table (\code{mapping_icd10_11}) to look for matches.
}
\examples{
icd10to11("a01")
icd10to11("a01", lang = "pt")
icd10to11("A06.6", simple = TRUE)
icd10to11("A01", simple = FALSE)
}
