% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd11_lookup.R
\name{icd11_lookup}
\alias{icd11_lookup}
\title{Look up ICD-11 code tables}
\usage{
icd11_lookup(x, simple = TRUE, lang = "en")
}
\arguments{
\item{x}{character. ICD-11 code for exact match.}

\item{simple}{logical. \code{TRUE} return a simplified result, with chapter, code and title (default). \code{FALSE} return the all variables.}

\item{lang}{character. Language for title. Default is \code{en}.}
}
\value{
a data.frame
}
\description{
Look up ICD-11 code tables
}
\examples{
icd11_lookup("1A03.1")
icd11_lookup("1A03.1", simple = FALSE)
icd11_lookup("1A03.1", lang = "pt")
}
